+++ ferron-modules-builtin/src/util/mod.rs
@@ -3,17 +3,11 @@ mod basic_auth;
 mod body_replacer;
 #[cfg(feature = "fcgi")]
 pub mod fcgi;
-#[cfg(any(feature = "rproxy", feature = "fauth"))]
-pub mod http_proxy;
-#[cfg(all(feature = "runtime-monoio", any(feature = "rproxy", feature = "fauth")))]
-mod send_net_io;
 #[cfg(any(feature = "dcompress", feature = "fcgi"))]
 mod split_stream_by_map;
 
 pub use basic_auth::*;
 #[cfg(feature = "replace")]
 pub use body_replacer::*;
-#[cfg(all(feature = "runtime-monoio", any(feature = "rproxy", feature = "fauth")))]
-pub use send_net_io::*;
 #[cfg(any(feature = "dcompress", feature = "fcgi"))]
 pub use split_stream_by_map::*;